+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Nothing yet.
 ### Changed
 - Now `php-hound` command return correct exit codes (0 = success, 1 = failure).
 
+### Fixed
+- If you use `git-diff`, it will only run tests over files referenced in the
+  diff.
+
 ## 0.7.0 - 2015-10-31
 🎃